refactor to support ILogger
